Убираю ненужное преобразование типа в макросе P7_LOG, добавленное по ошибке.

+++ Singleton/p7_logger.h
@@ -77,7 +77,7 @@ private:
 #ifdef USE_P7_LOG
 #define P7_LOG(log_level, module_enum, format, ...)\
    p7_logger_raii::instance().trace().P7_DELIVER(0,\
-                                            static_cast<size_t>(log_level),\
+                                            log_level,\
                                             p7_logger_raii::instance().module(static_cast<size_t>(module_enum)),\
                                             format,\
                                             ##__VA_ARGS__)
